requirements.txt
streamlit~=1.24.1
pandas~=1.5.3
numpy~=1.23.5
pivottablejs~=0.9.0
scipy~=1.10.1
scikit-learn~=1.2.2
Pillow~=9.5.0
pygwalker~=0.3.7
pandas-profiling~=3.6.6
ydata-profiling~=4.6.2
streamlit-pandas-profiling~=0.1.3
streamlit-toggle~=0.1.3
scikit-plot~=0.3.7
